requirements.txt
# Core Dependencies
requests==2.31.0
python-dotenv==1.0.0

# Oura Ring
# API client
requests-oauthlib==1.3.1

# Automated export (browser automation)
playwright==1.40.0

# Strava API
stravalib==1.6.0

# Google Sheets API
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.111.0

# Data Processing
pandas==2.1.4
numpy==1.26.2

# HTML Generation
jinja2==3.1.2

# Date/Time
pytz==2023.3

# Caching
diskcache==5.6.3
